requirements.txt
python-telegram-bot>=20.4
python-dotenv>=1.0.0
openai>=1.0.0